Refactor step3 to separate components
