+++ part2/countries/src/components/Filter.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const Filter = ({value, handler}) => {
+
+  return (
+    <>
+      <label>Search: </label>
+      <input value={value} onChange={handler}/>
+    </>
+  )
+}
+
+export default Filter
